1st Commit
HUB.TXT is the test-level-data for when the game initializes.
SPLASH.TXT is the splashscreen data for rendering the splashscreen when the game first runs.
SAVE.TXT is the user's initial save data. In it, it saves the latest location of the user in line #1 and the following lines record which gems were discovered by the user.

gemlord.bas is the BASIC language file for the game.